initial refactoring to service
