+++ tests/fixtures/services/bcl_convert_metrics_service/230622_A00621_0864_AHY7FFDRX2/Unaligned/Reports/Adapter_Metrics.csv
